BUG: repeated failing makes could generate bad output.
This is a quick fix to address two issues:

* Firstly, if the mec-tools mec_encrypt binary cannot be compiled,
gcc still produces an output file - it is a non executable "precompiled
header" file.  I assume there is a bug in the mec-tools makefile
logic, which I will look for later.

* Secondly, because I am using stdout for file generation, many commands
will still produce an output file even when they fail.  This output file
then looks "good" to the make command, which will carry on to do the
next step.

Since fixing the failed commands in my Makefile makes things more robust
in general, it also avoids the compile issue with the mec-tools.