Make the .txt.orig downloader more generic.
As two examples - When looking at the release notes for the x200 laptop,
they have the same "uc" name as the cdrom image.  The release notes for
the x270 laptop also has the same name as the cdrom image, but they use
neither the "uc" nor the "us" suffixes.

So, it seemed simplest to be the least smart and just download whatever
matches the cdrom image name.

This has the drawback that we sometimes end up with text files that tell
you how to use the windows .exe installer - but I am hoping that the
people reading these release notes will be able to realise that not all
instructions should be followed...