The L430 patch doesnt have all expected files.
It has used a different pattern for the patches applied to the rest
of the keyboard patches in this repository.  It may not even be able
to have the same exact sequence of patches.

It is not possible to test it with the standard testing framework,
but the only current way to exclude it from those tests is to mark
the text description with "Battery ..", which is not entirely
correct.

TODO: refactor the testing