Import trivial checksum helper.
This file was originally downloaded on each build from
    https://raw.githubusercontent.com/leecher1337/thinkpad-Lx30-ec/main/fwpat/util/npce885crc.c

To ensure no supply chain attacks were possible, a versioned method for
fetching this file was needed.  Since the containing repository was
quite large, a submodule was not ideal, but once the source was examined
and determined to be a small trivial checksum and not even as
complicated as a CRC generator, it seemed reasonable to just import the
file completely into this repo.

The version included here is the one from commit
cdb783ed2b7f8c18ac10d26ee668b358405973b6